refactor: reorganize SGP4 files into Vallado directory
