+++ include/AstCore/SGP4.hpp
@@ -1,0 +1,1 @@
+#include "../../src/AstCore/Propagator/SGP4/SGP4.hpp"
